Moved config into package
Moved all config into separate package
Everything is centralised, and loaded once per package
